requirements.txt
fastapi
uvicorn
pillow
requests
streamlit
torchvision
python-multipart